Uploaded some sprites
My SFML still isn't working with VS/VSCode, so I didn't want to push my whole repository. Manually uploaded sprites for Fire Hydrant and Tree obstacles, which I'll be using if/when I can get my engine working so I can do obstacle implementation. 
The dogRunner spritesheet is a WIP still, but the # of frames is final. So we should implement player animation based on this spritesheet when possible.